+++ CHANGELOG.md
@@ -5,6 +5,27 @@
 
 > All notable changes to this project will be documented in this file
 
+## [1.23.0](https://github.com/open-sauced/hot-sauce/compare/v1.22.0...v1.23.0) (2022-02-21)
+
+
+### 🍕 Features
+
+* load more records ([#127](https://github.com/open-sauced/hot-sauce/issues/127)) ([e6232f5](https://github.com/open-sauced/hot-sauce/commit/e6232f5386df29426e4c00d2d336d12d0b04ddf8)), closes [#81](https://github.com/open-sauced/hot-sauce/issues/81)
+
+## [1.22.0](https://github.com/open-sauced/hot-sauce/compare/v1.21.0...v1.22.0) (2022-02-21)
+
+
+### 🍕 Features
+
+* adds redirect for site_url ([#126](https://github.com/open-sauced/hot-sauce/issues/126)) ([1393e63](https://github.com/open-sauced/hot-sauce/commit/1393e632e9241aafa816634da23c466b5cc1a782)), closes [#108](https://github.com/open-sauced/hot-sauce/issues/108)
+
+## [1.21.0](https://github.com/open-sauced/hot-sauce/compare/v1.20.2...v1.21.0) (2022-02-21)
+
+
+### 🍕 Features
+
+* Add ability for a user to see the repos they have voted for ([#125](https://github.com/open-sauced/hot-sauce/issues/125)) ([8b9cefc](https://github.com/open-sauced/hot-sauce/commit/8b9cefcfde3487e98653869c4123bba238f87ca4)), closes [#124](https://github.com/open-sauced/hot-sauce/issues/124)
+
 ### [1.20.2](https://github.com/open-sauced/hot-sauce/compare/v1.20.1...v1.20.2) (2022-02-04)